PagesEditor: rename all .js files to .jsx
